INTERLOK-80 - Updated code to protect against API Changes INTERLOK-80 #time 30m
Because it's a nightly snapshot; the commons-csv devs have a tendency
to change the API without deprecating methods. Made the way in which
CustomFormatBuilder uses the withXXX methods, to be reflection based so
that we can be "protected" at compile time; however if they change the
API massively one of the tests will start failing.