INTERLOK-3424 Implement aggregate method from interface.
Directly implement aggregate() method that is now defined in
MessageAggregator. joinMessage() delegates to it
Add support for filtering which was missing in the original implementation
since it's down to the impl to figure that out.
Add support for overwriting metadata which was missing in the original
since it's down to the impl to figure that out.